requirements.txt
websockets
openai
python-dotenv
numpy
